Update project to 0.5.0

+++ adpcm_codec.h
@@ -1,4 +1,4 @@
-/* SPDX-FileCopyrightText: Copyright 2024 Azamat H. Hackimov <azamat.hackimov@gmail.com> */
+/* SPDX-FileCopyrightText: Copyright 2025 Azamat H. Hackimov <azamat.hackimov@gmail.com> */
 /* SPDX-License-Identifier: LGPL-2.1-or-later */
 #pragma once
 
@@ -9,7 +9,7 @@
 typedef struct ADPCMChannelStatus {
   int predictor;
   int16_t step_index;
-  int16_t prev_sample; // for encoding
+  int32_t prev_sample; // for encoding
 } ADPCMChannelStatus;
 
 int adpcm_rib_decode_frame(const std::shared_ptr<std::vector<int8_t>> &in_stream,
